converted modifier to standard code structure we’re using
